chore : layout 구조 변경, Register page client 분리
