+++ src/app/(site)/community/[id]/ContentsDetail.tsx
