+++ src/app/(site)/community/[id]/page.tsx
