+++ src/app/(site)/layout.tsx
@@ -1,0 +1,14 @@
+import Header from '@/components/common/Header';
+
+export default function SiteLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <>
+      <Header />
+      <main>{children}</main>
+    </>
+  );
+}
